redid books to match style guide; fixed treaty date format
Threw together a (messy) version of book rendering to bring citations in line with the McGill guidelines on ordering and formatting of editors, translators, volume/series titles, and (especially) volume numbers (per ss. 6.2.1–6.2.2 & 6.2.4–6.2.5).
There's definitely a more efficient way to write this but I wasn't in the headspace for it, so if anyone else wants to fork this and take a swing I encourage it; else I'll maybe get around to it at some point (but unlikely, seeing as it works as is).

Also updated treaty formatting to better fit with s. 5.1.1, correcting the way that dates are rendered and adding short titles.